Don't commit transaction in output_layout_handle_change
Because this is only ever emitted by wlroots in response to a direct
function call by us (such as in apply_output_config) or in response to
an output event such as a mode event, and we already commit transactions
in the output event handlers.

This prevents creating a transaction while the tree is in an
inconsistent state, which caused a crash.