config: allow for inline to end of line comments
Allows for the following syntax in the config:
    `command <arguments...> # the rest of this line is a comment`
Example:
    `input * repeat_delay 200 # milliseconds`

Note: The space before and after the `#` is mandatory otherwise the
`#` will be seen as part of an argument (such as colors). If a `#`
is needed as it's own argument, then it can be enclosed in quotes. For
example, `bar bar-0 separator_symbol "#"`.