Added null pointer checks to non-static functions
Since create_list can return NULL, functions receiving a list
should check for null pointers I believe.
Lists are used a lot throughout sway and often are assumed not to be
NULL. It's a small thing but I haven't been using sway for that long
and I figured this is a good place to start.