Add swayrun command to run sway using a login shell
When running Sway from a display manager, a user's login shell is
not used. Because of this, profile scripts are not run and the
environment may not be configured. As suggested in #3109, this can
be solved by a wrapper that runs sway via the user's configured
shell. This is similar to the approach taken by GNOME to source
the standard set of profile scripts.

This could be accomplished with a shell script, but a dedicated
binary avoids dependencies on specific shells.

The sway.desktop wayland session file is also updated to call the
swayrun wrapper.