swaymsg pretty-print for get_bindings
This adds the capability to swaymsg to pretty-print the bindings returned
by the get_bindings IPC command.

Each line of the pretty output contains a text representation of the
modifiers and key symbols along with the bound sway command.

The sort order of the bindings attempts to group bindings first by common
command prefix and then by key modifiers and symbols. This hybrid sort is
meant to be useful to a human reader.

Signed-off-by: Peter Grayson <pete@jpgrayson.net>